Move cluster code from lib/imagery to lib/cluster Rename remaining imagery library from libgrass_I to libgrass_imagery
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@31744 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ lib/cluster/c_assign.c
@@ -1,6 +1,5 @@
 #include <math.h>
-
-#include <grass/imagery.h>
+#include <grass/cluster.h>
 
 int I_cluster_assign ( struct Cluster *C, int *interrupted)
 {
@@ -25,7 +24,7 @@ fprintf (stderr,"I_cluster_assign(npoints=%d,nclasses=%d,nbands=%d)\n",
 	    d = 0.0;
 	    for (band = 0; band < C->nbands; band++)
 	    {
-		q = (double) C->points[band][p];
+		q = C->points[band][p];
 		q -= C->mean[band][c];
 		d += q*q;
 	    }
@@ -38,7 +37,7 @@ fprintf (stderr,"I_cluster_assign(npoints=%d,nclasses=%d,nbands=%d)\n",
 	C->class[p] = class;
 	C->count[class]++;
 	for (band = 0; band < C->nbands; band++)
-	    C->sum[band][class] += (double) C->points[band][p] ;
+	    C->sum[band][class] += C->points[band][p] ;
     }
 
 	return 0;
